% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfs.R
\name{box_output}
\alias{box_output}
\title{Generating readable image output}
\usage{
box_output(table, border)
}
\arguments{
\item{table}{data table}

\item{border}{Colour used to surround data table}
}
\value{
image of data table with coloured border
}
\description{
Generating readable image output
}
