% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfs.R
\name{cfa.tab.multi}
\alias{cfa.tab.multi}
\title{Generating Concatenated APA-formatted tables of results}
\usage{
cfa.tab.multi(x, y, z = NULL, a = NULL, b = NULL, robust = FALSE)
}
\arguments{
\item{x, y, z, a, b}{lavaan model object(s)}

\item{robust}{Whether robust RMSEAs are calculated. Defaults to FALSE.}
}
\value{
data.frame object in APA format
}
\description{
Generating Concatenated APA-formatted tables of results
}
