% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfs.R
\name{cfa.tab}
\alias{cfa.tab}
\title{Generating APA-formatted tables of results}
\usage{
cfa.tab(x, robust = FALSE)
}
\arguments{
\item{x}{lavaan model object}

\item{robust}{Whether robust RMSEAs are calculated. Defaults to FALSE.}
}
\value{
data.frame object in APA format
}
\description{
Generating APA-formatted tables of results
}
