% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfs-package.R, R/package-package.R
\docType{package}
\name{elfs-package}
\alias{elfs-package}
\title{elfs: Evaluating and Estimating Latent Factor Scores}
\description{
What the package does (one paragraph).

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Neil Hester \email{neil.hester@uwaterloo.ca}

Authors:
\itemize{
  \item Neil Hester \email{neil.hester@uwaterloo.ca}
  \item Shawn Yee \email{s6yee@uwaterloo.ca}
}

}
\keyword{internal}
