% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfs.R
\name{elfs}
\alias{elfs}
\title{Estimates and evaluates latent factor scores for scales}
\usage{
elfs(
  data,
  f1_cols,
  f2_cols = NULL,
  f3_cols = NULL,
  f4_cols = NULL,
  f5_cols = NULL,
  f6_cols = NULL,
  fg_cols = NULL,
  ordered = FALSE,
  missing = "listwise",
  dynamic = FALSE,
  meas_invar = NULL,
  modify = NULL,
  lfs_method = NULL,
  lfs_transform = TRUE,
  chrome_bypass = FALSE,
  f1_name = NULL,
  f2_name = NULL,
  f3_name = NULL,
  f4_name = NULL,
  f5_name = NULL,
  f6_name = NULL,
  fg_name = NULL
)
}
\arguments{
\item{data}{data.frame object}

\item{f1_cols, f2_cols, f3_cols, f4_cols, f5_cols, f6_cols, fg_cols}{character vector(s) listing column names included in (each) latent factor. Only \code{f1_cols} is required.}

\item{ordered}{whether to treat measured variables as ordinal (vs. continuous; see \code{\link[lavaan:sem]{lavaan::sem()}})}

\item{f1_name, f2_name, f3_name, f4_name, f5_name, f6_name, fg_name}{optional names for each specified latent factor.}
}
\description{
Estimates and evaluates latent factor scores for scales
}
