% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfs.R
\name{h}
\alias{h}
\title{Calculating Coefficient H}
\usage{
h(model)
}
\arguments{
\item{model}{lavaan fitted model object}
}
\value{
numeric indicating Coefficient H
}
\description{
adapted from \if{html}{\out{<a href = "https://github.com/jsakaluk/scripty">}}jsakaluk/psyscores\if{html}{\out{</a>}}
}
