% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfs.R
\name{head_output}
\alias{head_output}
\title{Generating readable output}
\usage{
head_output(table, color)
}
\arguments{
\item{table}{data table}

\item{color}{Colour used to surround data table}
}
\value{
data table with coloured border
}
\description{
Generating readable output
}
