% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfs.R
\name{lav_string_tau}
\alias{lav_string_tau}
\title{Generating lavaan-readable tau-equivalent single-factor string from model specifications}
\usage{
lav_string_tau(lfs_name, lfs_cols)
}
\arguments{
\item{lfs_name}{Name of latent factor}

\item{lfs_cols}{Column(s) loading onto latent factor}
}
\value{
single-factor string readable by lavaan
}
\description{
Generating lavaan-readable tau-equivalent single-factor string from model specifications
}
