% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortDarkTriad_data.R
\docType{data}
\name{ShortDarkTriad_data}
\alias{ShortDarkTriad_data}
\title{Short Dark Triad (SD3) Dataset}
\format{
\subsection{\code{ShortDarkTriad_data}}{

A data frame with 17,738 rows and 29 columns:
\describe{
\item{M1, M2, M3, M4, M5, M6, M6, M7, M8, M9}{Participant responses for items measuring Machiavellianism (range: 1-5)}
\item{N1, N2, N3, N4, N5, N6, N6, N7, N8, N9}{Participant responses for items measuring Narcissism (range: 1-5)}
\item{P1, P2, P3, P4, P5, P6, P6, P7, P8, P9}{Participant responses for items measuring Psychopathy (range: 1-5)}
\item{country}{Two-letter code for each participant}
\item{source}{How the user came to take the test. 1 = from front page of website, 2 = Google search, 3 = other (based on HTTP referrer)}
}
}
}
\source{
\url{https://openpsychometrics.org/_rawdata/}
}
\usage{
ShortDarkTriad_data
}
\description{
Responses to the Short Dark Triad (SD3; Paulhus & Jones, 2001) collected from
the Open-Source Psychometrics Project. Responses were filtered to retain complete data,
after converting responses of 0 to NA.
}
\keyword{datasets}
