% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfs-package.R
\docType{package}
\name{elfs-package}
\alias{elfs-package}
\title{elfs: Evaluating and Estimating Latent Factor Scores}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Automated evaluation/estimation of latent factor scores, building on 'lavaan'. Provide the data in wide format, column names for each factor, and receive comprehensive information regarding model fit, reliability, factor correlations, measurement invariance, and dynamic fit indices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SarcylY/elfs}
  \item \url{https://sarcyly.github.io/elfs/}
  \item Report bugs at \url{https://github.com/SarcylY/elfs/issues}
}

}
\author{
\strong{Maintainer}: XXX XXX \email{redacted@gmail.com}

Authors:
\itemize{
  \item XXX XXX \email{redacted@gmail.com}
  \item YYY YYY \email{redacted@gmail.com}
}

}
\keyword{internal}
